backend/vision/requirements.txt
absl-py==1.0.0
agate==1.7.1
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.9
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.12.1
backcall==0.2.0
bcrypt==4.1.2
beautifulsoup4==4.11.1
bidict==0.23.1
bleach==3.3.0
boto3==1.24.7
botocore==1.27.7
build==1.2.1
cachetools==5.0.0
certifi==2021.5.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colorlog==6.8.2
contourpy==1.0.6
coverage==5.5
cryptography==42.0.5
csvkit==1.1.1
cycler==0.11.0
Cython==0.29.32
dbfread==2.0.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
dnspython==2.3.0
docopt==0.6.2
docutils==0.17.1
ecdsa==0.19.0
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
fastapi==0.110.0
fastjsonschema==2.16.2
filelock==3.14.0
flake8==7.0.0
Flask==2.0.3
flatbuffers==22.12.6
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
future==0.18.2
gast==0.4.0
geojson==2.5.0
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googlemaps==4.10.0
grpcio==1.44.0
h11==0.14.0
h5py==3.6.0
haversine==2.3.1
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.5.1
hypothesis==6.14.0
idna==3.6
iniconfig==1.1.1
ipykernel==6.17.1
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.3
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.2.0
json5==0.9.14
jsonpointer==2.3
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter_server==2.7.3
jupyter_server_terminals==0.4.2
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
jupyterlab_server==2.24.0
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
leather==0.3.4
libclang==13.0.0
lxml==4.9.2
mapbox-vector-tile==1.2.1
mapillary==1.0.12
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mercantile==1.2.1
mistune==2.0.4
multidict==6.0.5
munch==2.5.0
mysql-connector-python==8.3.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.16.4
nbformat==5.7.0
nest-asyncio==1.5.6
netifaces==0.10.6
notebook==7.0.3
notebook_shim==0.2.2
numpy==1.21.0
oauthlib==3.1.1
olefile==0.46
openai==1.33.0
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.9.15
outcome==1.2.0
overrides==7.4.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parsedatetime==2.6
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickle5==0.0.11
pickleshare==0.7.5
Pillow==9.0.0
pip-chill==1.0.3
pip-tools==7.4.1
pipreqs==0.5.0
pkginfo==1.7.0
platformdirs==4.2.1
pluggy==0.13.1
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pyasn1==0.6.0
pyasn1-modules==0.2.8
pyclipper==1.3.0.post5
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.16.3
pyflakes==3.2.0
Pygments==2.9.0
pymongo==4.3.3
PyMySQL==1.0.2
pyotp==2.6.0
pyparsing==2.4.7
pyproject_hooks==1.1.0
pyrsistent==0.19.2
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-jose==3.3.0
python-json-logger==2.0.4
python-multipart==0.0.9
python-slugify==8.0.1
python-socketio==5.11.2
pytimeparse==1.1.8
pytz==2021.3
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
readme-renderer==36.0
referencing==0.30.2
regex==2022.3.15
requests==2.31.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
robin-stocks==2.1.0
rpds-py==0.10.2
rsa==4.9
s3transfer==0.6.0
sacremoses==0.0.49
scikit-learn==1.2.2
scipy==1.7.3
seaborn==0.12.1
selenium==4.10.0
Send2Trash==1.8.2
Shapely==1.8.1
simfin==0.9.1
simple-websocket==1.0.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.1
socketio==0.2.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.47
stack-data==0.6.1
starlette==0.36.3
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-macos==2.9.2
termcolor==1.1.0
terminado==0.17.1
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.11.6
toml==0.10.2
tomli==2.0.1
torch==1.11.0
tornado==6.2
tqdm==4.61.1
traitlets==5.7.0
transformers==4.18.0
trio==0.22.0
trio-websocket==0.9.2
turfpy==0.0.7
tweepy==4.3.0
twine==1.13.0
typing_extensions==4.10.0
tzdata==2023.3
ujson==5.9.0
undetected-chromedriver==3.5.0
uri-template==1.2.0
urllib3==1.26.5
uvicorn==0.29.0
uvloop==0.19.0
vaderSentiment==3.3.2
virtualenv==20.26.1
vt2geojson==0.2.1
watchfiles==0.21.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==12.0
Werkzeug==2.0.3
widgetsnbextension==4.0.4
wrapt==1.14.0
wsproto==1.2.0
xlrd==2.0.1
yarg==0.1.9
yarl==1.9.4
